example/step_functions/workflow-init/requirements.txt
python-dotenv==0.17.1
boto3==1.17.44